Move syscall implementations to libsyscalls.
This replaces libc.a and libc++.a with linker scripts to keep existing code
linking.  Other options are possible for libc, but libc++.a must be a linker
script because the compiler ensures that it comes after any libraries
provided on the command line.

This is a work in progress.  In particular:
 - The current set of symbol maps needs some cleaning up and auditing.
 - i386 is broken (and thus amd64).