Stop tracing when the currently traced program crashes
When QEMU per-thread tracing is enabled and the traced program is killed
due to an exception the trace would previously continue until the next
context switch because the MDTD_QTRACE was only being checked in swtch.S

When using qtrace this would result in the trace running until qtrace
terminates. As we are not interested in those instructions if we have per-
thread tracing enabled just stop the trace as soon as a CHERI exception is
logged.