Don't cast to __capability in __DECONST/__DEVOLATILE
With the cast to a __capability type we will get a warning in hybrid
when casting from a plain pointer to a capability and it might also add
unnecessary instructions if the compiler can't optimize the cfromptr
away. For hybrid mode I added __DECONST_CAP, etc. macros to dequalify
capabilities. So far it is not being used but might be in the future.