Further updates to libcheri to support the new $csp-based ("nosp") interpretation of the stack capability:
- Modify libcheri to generate $csp-based stack pointers on sandbox creation.
- Modify the shared-memory interface with the sandbox to offer a $csp-based
  stack pointer for use in the sandbox.
- Adjust the invocation vector inside the sandbox to expect a $csp-based
  stack pointer.
- Adjust the system-class invocation vector outside of a sandbox to expect
  a $csp-based stack pointer in CheriABI.
- Adjust hybrid and pure-capability invocation functions to manage the stack
  using $csp semantics.

With these changes, pure-capability code runs fine within a sandbox, and the
hybrid-capability cheritest tool passes all expected tests invoking
sandboxes.

Due to a possible code-generation bug in the "nosp" compiler branch,
cheriabitest is not yet fully functional (although it does partially run),
due to unexpected use of $c28 in userspace, which is a kernel-only register.