Update rtld to work with 'nosp'.
Most of this is just using zero instead of sp with $c11 and using
cincoffset to reserve the local stack frame.  However, rtld_start was
previously reserving stack space in two stages and I've changed it to
reserve it all at once in the beginning.  This required adjusting some
of the constant offsets relative to $c11.  One 'dla' that was used to
load a constant into a register was also changed to 'dli'.