Rework the syscall calling convention.
Rather than requesting a system call by number, use the syscall
number to retieve a token (capability) from __systokens[] and pass
that to the kernel.  __systokens[] is a pointer to a kernel-provided
array of tokens whose location is passed in the AT_SYSTOKENS element
of auxargs.

Tokens are sealed capablities with an offset equal to the system
call number making them unforgable.  By using multiple object type
or different base addresses, we should be able to do interesting
things like restricting a file descriptor to a particular sandbox.

The kernel currently supports old binaries, but new binaries will
not work with old kernels (this could be done, but seems not worth
the trouble on CheriBSD).