Fix time being twice too slow under Qemu.
This code makes it possible to make the kernel automatically take
the CCRes register into account.  It's enabled by passing -1 as
the last argument to mips_timer_init_params.  New sysctl,
machdep.counter_ccres, allows for querying the current value.

For now it's only enabled for MALTA; it depends on rdhwr and might
not work for some older MIPS versions.