Add alignment support to __libc_allocate_tls().
Pass the greater of TLS_TCB_ALIGN or the alignment of the TLS init
segment as the alignment in _init_tls.

The current implementation is inefficent in that it uses the
malloc_aligned() from rtld to align the whole allocation and then pads it
as required to align the TLS region to the minimum alignment.  This
results in a significant overallocation.

Also, tidy up support for tcbsize != TLS_TCB_SIZE.  I suspect it remains
wrong, but it is (hopefully) somewhat less confusing.