Add file capablity versions of most syscalls that take file descriptors.
This implementation uses a pair of functions (fc2fd() and fd2fc())
to convert between file capabilities and file descriptors in the
entry path (in the common case) and in the return path for syscalls
that return descriptors.  The latter set of APIs have been modified to
return file capabilities by reference.

This implementation avoids deep changes to file handle use at the
cost of a race between fc2fd() calls and the actual fget() etc to
obtain the file pointer.  During the window the file could be changed.
A similar race exists in the audit code.  Both could be closed by making
the capabilty the kernel handle to files and resolving legacy file
descriptors to capabilities before auditing.