Begin to flesh out rtld constructor/destructor and invocation failure-mode handling: when something goes wrong (e.g., trusted-stack overflow), clear registers to ensure that we're not leaking trampoline state (e.g., a capability to the trusted stack!) back to the caller, and also set $a0 to report the error.  More refinement remains required, including selecting suitable error codes, and in error checking during domain switching.
Unlike in the kernel case, this currently focuses on synchronous error
reporting to the caller, rather than signal delivery.  But perhaps a
signal would be prefered for (e.g.,) local-global violations.

Leave "return" error handling the same: generate a trap on a trusted-stack
underflow, although a signal might be more desirable in the future.