Allow linking the CHERI_MALTA kernels with LLD
MALTA64 was already working as expected but when linking a CHERI_MALTA
kernel it was dying with the following error during boot:
```
tcp_init: net.inet.tcp.tcbhashsize auto tuned to 32768
Trap cause = 4 (address error (load or I-fetch) - kernel mode)
[ thread pid 0 tid 100000 ]
Stopped at      cpu_switch+0xd0:        ld      t2,0(t2)
```

It turns out this happens on the first context switch when we check the
value of `qemu_trace_perthread`. LLD assigns a different address to it:

BFD: ffffffff807901b8 g       .sbss           00000004 qemu_trace_perthread
LLD: ffffffff807c481c g       .bss            00000004 qemu_trace_perthread

As we can see here it is only aligned to four bytes now (which is perfectly
fine because it is declared as an int). The trivial fix is to stop using
PTR_LA for an int variable.

Note: I wanted to use the INT_LA macro but it seems that it is defined to
`ld` for MIPS64 which looks wrong to me.