Marshall the kevent uap to g_kevent_args for CPU_CHERI.
kib@'s recent change to add abstime timestamps for kevent() added a
new 'g_kevent_args' which is like the auto-generated 'struct kevent_args'
without the explicit padding in sysproto.h.  The sys_kevent() system
call then casts the 'uap' to 'struct g_kevent_args *' when calling
kern_kevent_generic.  For the CHERI case this causes the uap to
be treated as if it did not have the padding in place to pad all
arguments out to uintcap_t resulting in kevent() misparsing its arguments
and failing (I got EFAULT errors under 128 and EBADF under 256 trying
to mount NFS).

Fix this by allocating a 'struct g_kevent_args' on the stack in the
CPU_CHERI case and populating it with members of 'uap'.