Attempt 2 at making sandbox.ld work with LLD
BFD implcitly adds any .o file mentioned in the linker script to the
list of input files whereas LLD only treats the file name as a pattern
and needs an explicit INPUT() for the command to work.

LLD is now creating the correct output sections but it appears that
they are empty because it appears to hardcode that .text input sections
must go in the .text output section.