Trick LLD into placing rtld_vector and invoke vector into separate sections
For some reason LLD is ignoring the linker script directive to place
either .text or even the * wilcard from crt_rtld.o into
rtld_vector_text. Work around this by changing the assembly to emit the
code into rtld_vector_text/invoke_vector_text directly.

I will see if this is also an error in the latest version of upstream LLD
and either do a merge to that version if it is fixed or report a bug.

The created binary now looks correct and has the rtld/invoke code at
the correct virtual addresses so hopefully it will finally work now.