Add new type vm_ptr_t intended as a replacement for vm_offset_t to handle possibly different semantics of vm_offset_t.
Currently vm_offset_t is used as an integer virtual address in some address space, in principle the virtual address is not a pointer but it is sometimes used to create one.
This is somewhat ambiguous as vm_offset_t is not necessarily a reference to any object but is sometimes used as one.
The vm_ptr_t type represents a dereferenceable object returned by the pmap and vm modules, it is not an integral virtual address.
The usefulness of this separation remains to be seen but can relatively easily be reverted in the future.