Implement __CAP_CHECK(cap, len).
This macro verifies that a capability offset (and 0 or more bytes after
it) are inbounds.

Add a help macro __DECAP_CHECK(cap, len) which checks that the given
length is inbounds and returns the virtual address if so (NULL if not).

Replace many __cheri_fromcap casts with __DECAP_CHECK().

__DECAP_CHECK() should be used with care and should only be used where
the capability in question could be from user-space.   Where it is not
(e.g. the UIO_SYSSPACE case when handling a struct uio) __cheri_fromcap
casts should be used as they pose no risk of NULL confusion.

Add a __CAP_CHECK() to verify bounds.

Implement __DECAP_CHECK using __CAP_CHECK().