Change rwlock so that we cast to integers the flags extracted from the rw_lock owner pointer field.
This is necessary otherwise the compiler will use capability branches and comparisons with canonical NULL instead.
The change is justified since the flags are meant to be integers without base or bound and they must be extracted as such,
embedding the flags in the pointer is not changed as the pointer offset is being changed as intended and we have to
preserve all the other fields.

Currently there are some new macros RW_GET_<flag_name>(rw->rw_lock) that perform the bitwise AND on the integer cast of
the pointer, these have been replaced in all places where flag values are compared. This seems a better approach than
inserting casts in the same places although it might impact readability.