Make helloworld work again by not setting bounds on PCC
Otherwise rtld will crash when calling the init function for libc.so and
all the cgetpccsetoffset sequences used for cross-library calls will
also fault. For static binaries we still set the bounds of PCC to be the
end of the text segment. In the future (cap-table) we will set bounds on
PCC for dynamically linked executablesi, but right now this is not
possible.