rtld-cheri-elf: let the kernel load relocbase into $c4
The kernel has that information easily available whereas it is quite hard
to do this in the assembly in a way that LLD will accept.
LLD refuses insert an absolute reloction into PIC binaries which is the
approach currently used to calculate the image base address. We could
modify lld to provide a symbol that can be loaded using PIC_LOAD_CODE_PTR
but as we already have the information available that seems overly
complicated.

This exposed some issues in _rtld_relocate_nonplt_self(): it was casting
relocbase to uintcap_t and adding that to all the got entries. In the pure
capability ABI this only adds the offset and not the base address.
Previously the base address was 0 and the offset was relocbase, but now
that the kernel sets relocbase to be the load address of rtld with bounds
up to the end this no longer works. Fixed by casting to vaddr_t instead.

This is almost the last change required before we can finally stop using BFD.