Stop using the hacky PIC calls in rtld-cheri-elf assembly
They only work with BFD since it will always fill in a value relative to
the load address in the dla relocations. LLD refuses to link a program with
relocations that are position dependent unless -Wl,-z,notext is passed.
However, these are not quite the same since LLD will actually generate
dynamic relocations (in .rel.dyn) instead of just filling in the values.