Change virtual memory initialization, vm_map and vmem to support allocations using capabilities.
vm_map now requires a valid capability for the mappable region to be provided when the map is created, vm_map should use
the capability to restrict the mappings that it provides. Currently this only happens for vm_map_find().
The kernel_map is initialized using capabilities initialized in pmap_bootstrap() in XKSEG (cheri-mips).
The vmem allocator receives bounded mappings for the kernel_arena and should properly return capabilities from vmem_alloc().