Fix mallocinit initialization of the mt_zone.
The same UMA debug mtrash_{ctor, dtor, init, fini} are used for both the mt_zone and kmemzones although mt_zone items are not changed to store
the related malloc_type* in the last pointer-sized block of memory in the allocated item so that it can be recovered for debugging.
Use normal uma trash constructors in this case (when INVARIANTS are enabled).