Do not run normal mtrash_{ctor, dtor} when building for CHERI-256 pure capability kernel.
The smallest kmemzone items are 16 bytes, which are not enough to accomodate a struct malloc_type* that is stored in the free'd item when INVARIANTS are enabled.
In this case we just run the default trash_{ctor, dtor} that give no information about the previous owner of the allocation but still detect use-after-free.