Change the interface of kmem_malloc and kmem_back to use vm_ptr_t instead of vm_offset_t.
This change is minimally required to have a working malloc. The vm and kmem interface needs more refinement.
In general it is desirable to require valid capabilities instead of virtual addresses for all external functions in vm so that
we can refuse to manipulate pages for arbitrary virtual addresses.
In principle all kmem and kernel virtual addresses should come from the kernel_map which is the only part of vm that should
be allowed to hand over capabilities for the kernel mapped AS region.