Port MipsUserGenException to purecap kernel.
This changes various macros to accomodate changes in the way the stack pointer is saved and restored.
Some load and stores to PCB are also tricky because in cheri256 the offset field in cld/csd is too small
to fit the immediate offset in PCB for some registers. I did not optimize out the few additional loads
that are not required in cheri128 yet.