Port SI_SUB_KLD sysinits to purecap kernel.
link_elf_init for link_elf is changed to generate capabilities with correct bounds for pointers in the module elf files, including the kernel.
The elf_file.address and linker_file.address are now capabilities that span the elf image loaded in memory for the kernel and kernel modules. The former is used to generate other capabilities in the elf image.
Currently a global capability is used to generate pointers in the kernel elf_file because kernels are linked at a specific nonzero address while modules are linked 0-relative, this causes contention in how the relocation address is applied almost everywhere.