Make most __mac_* syscalls take capability arguments.
Split user and kernel represenations of struct mac and use a kmac_t
typedef to hide the choice on each architecture.

Add new socket option directions SOPT_GETCAP and SOPT_SETCAP and use
them to signal to sooptcopyin() that capabilities should be copied.  No
support in sooptcopyout() as the only uses were in gratuitous copyouts
in the getsockopt path (the struct mac is used to find the data to
touch, but is not altered).