Stop mapping all ELF segments as writable
Now that we process __cap_relocs in LLD it is an error to have a relocation
point to a read-only section so there should be no more crashes
during crt_init_globals() now.

This reverts the HACK from 2015 (4b015538) which was updated for the
removal of EM_MIPS_CHERI in 2017 (45ce2e09). It took a while but we
should now finally be able to run CHERI binaries with actual read-only
segments.