Fix purecap kernel proc0 initialization.
When creating vmspace0 vm_map we need to give valid capabilities in xuseg for that start and end of the map.
We use the sysentvec entries sv_minuser and sv_maxuser to bound the map in xuseg.
This should probably apply to imgact in a similar way but has not yet been done.