Handle R_MIPS_64 relocation in RTLD
Apparently plain MIPS code never generated a R_MIPS_64 relocation but had
everything as relative ones. However in the __cap_relocs section we really
want an absolute value to be written and not just the load address.
This was previously causing crashes in __libc_interposing_table since all
the capabilities were just the load address with the correct size so we
were trying to interpret the ELF header as instructions.

I added a fix to LLD to emit R_MIPS_64 instead of R_MIPS_REL32 but RTLD
didn't understand dynamic R_MIPS_64 relocs so I now added this. I believe
we should also add this to the plain MIPS RTLD (glibc supports it too, so
there must be some plain MIPS binaries that need it).