rtld: Accept R_MIPS_CHERI_ABSPTR relocations to write absolute values (without using a GOT slot)
Turns out we can't just use R_MIPS_64. Glibc claims the following:
```
For full compliance with the ELF64 ABI, one must precede the
_REL32/_64 pair of relocations with a _64 relocation, such
that the in-place addend is read as a 64-bit value.
```

Let's avoid any kind of conflict with tools that follow that ABI and use
a different type instead.