rtld-cheri-elf: make_function_pointer should only sometimes restrict bounds
This is needed in order to support older CheriABI binaries. For now I added
two flags to the rtld object structure:

If `restrict_pcc_basic` is true we can restricted $pcc to the object's
executable segment, if `restrict_pcc_strict` is true we can restrict it
to just the functions bounds in dlsym()/dlfunc(). Otherwise we need to
fall back to giving the full address space bounds in $pcc.