Handle differently sized struct ifreq's in ioctl definitions.
Replace "case <CMD_USING_IFREQ>:" with a macro
"CASE_IOC_IFREQ(<CMD_USING_IFREQ>):".  In the default case the macro
expands to the former.  In a CHERI kernel, it expands to the origional
case statement and one with the command size adjusted for CheriABI.

Replace direct comparisons in if statements or ternary operations with
switch statements.  This involved manual reformatting and some unrolling
of combined switch statements where the paths were too dissimilar.

The first set of changes was accomplished automatically with the script
fix-ifreq-ioctl-switches.sh.  The second set was manual, but aided by
the script find-ifreq-non-switch.sh.