Add 32-bit compat for ioctls that take struct ifgroupreq.
Use an accessor to access ifgr_group and ifgr_groups.

Use an macro CASE_IOC_IFGROUPREQ(cmd) in place of case statements such
as "case SIOCAIFGROUP:".  This avoids poluting the switch statements
with large numbers of #ifdefs.

Clean up comments and avoid a gratutious assigment to a temporary of
the same type.