Refactor PCIOCGETCONF for improved readability.
The code now has a single, consistant flow for all three ioctl
variants.  ifdefs and for pre-FreeBSD-7 compatability are moved to
functions and macros.  So the flow is alwasy the same, we impose
the cost of allocating, copying to, updating from, and freeing a
copy of struct pci_conf_io on all paths.

This change will allow PCIOCGETCONF32 support currently in
sys/compat/freebsd32/freebsd32_ioctl.c to be moved here.