Stop using fuswintr() and suswintr().
Always take the AST path rather than calling a MD functions which are
often implemented as always failing.  The is the case on amd64, arm,
i386, and powerpc.  This optimization (inherited from 4.4 Lite) is a
pesimization on those architectures and is the sole use of these
functions.  They will be removed in a seperate commit.