Make -Wcheri-bitwise-operations warnings an error by default
This warning is quite noisy and has a lot of false positives but for the
other cases it will result in very hard to find bugs (such as locks not
being released and a deadlock much later on). Now that I've fixed the
remaining warnings I believe we should try to keep it this way and have
this be a -Werror warning.