Avoid printing "dlopen: not available when statically linked" under kyua
This message breaks a bunch of tests since they assume that stderr will be
empty. I find this message useful since it helps me find errors when we
build a normally dynamic binary statically linked (e.g. postgres) so I only
disable it when running under kyua. It appears that the easiest way to
detect this is to check whether the magic env var __RUNNING_INSIDE_ATF_RUN
is set.