Fix casueword{,32}_c zeroing destination
The second argument to Capability Store Conditional instructions is the
value to store, not an offset, with the first one being where the
boolean result is placed. Since we can do store conditionals without
clobbering the source, unlike standard MIPS, we also no longer need to
make a copy of the new value input and can just use a2 directly.

This was causing contested locks to have their owner cleared if subject
to unfortunate timings, as casueword32_c is used to set the
UMUTEX_CONTESTED flag on the owner when performing a UMTX_OP_MUTEX_WAIT.