Actually enable capsicum syscalls for CheriABI.
This lead to a weird failure in diff if one of the files on the
command line was "-".  In that case, diff calls caph_limit_stdio()
which places sensiable limits on stdin, stdout, and stderr.  It
ignores ENOSYS in all cases and EBADF descriptor is closed, but
only does the altter if cap_rights_limit() returns EBADF.  Due to
the error this commit fixes, cap_rights_limit() and cap_ioctls_limit()
both returned ENOSYS because the wrong CheriABI implementation was
selected.  cap_fcntls_limit() doesn't take pointer arguments and thus
was enabled and returned EBADF on fd=2 which caused an error to be
returend and diff to fail.