Fix kernel stack size increase for cheri256.
The stack in cheri purecap uses a reserved TLB entry that maps two contiguous 16K pages.
In hybrid CheriBSD, the lower of the two pages, is a guard page.
Due to stack exhaustion in cheri256 purecap kernel, we remove the guard page and we use both 16K pages for
the kernel stack (and PCB).
Removing the guard page is safe because the stack and PCB are properly bounded.

Currently this is only active for cheri256 but if it works well we can enable it for cheri128 as well, since
the guard page should not be needed in that case as well.