Turn unaligned access emulation off by default
It currently doesn't handle any of the cheri branches + cjr/cjalr which
means an unaligned trap in a branch delay slot there will silently jump to
the wrong address if the branch is actually taken. This then results in
really weird faults later on. Until the emulation is actually performed
correctly it is probably better to have the process die where the problem
is rather than continuing. Emulating CCall due to an unaligned trap also
seems like a bad idea...

Furthermore it seems like the branch emulation is wrong for at least some
of the MIPS branch-likely instructions since not all of them have delay
slots.