Include /bin/sh (but without a hardlink) in cheribsdbox
This will be useful if the environment is completely messed up (e.g. due
to copying a broken rtld-cheri-elf.so to the rootfs). However, we can't use
cheribsd as /bin/sh since it won't work with su:
su will invoke /bin/sh with argv[0] as _su so cheribsdbox complains that it
can't run program _su.