Include all the SSH binaries in cheribsdbox (built without kerberos)
This allows us to remove all the kerberos shared libs from the image and
should shrink the image sizes noticeably. I was able to login with a
cheribsdbox sshd so it does appear to work.

I'm not sure if it makes sense to build a separate binary for all the ssh
stuff but for now this seems to work just fine.