Reorganise the busdma types to work with CHERI.
The following types are changed:
bus_addr_t: address in the bus space, this is not dereferenceable as a virtual address
	   but is 'dereferenceable' in the bus space (e.g. this would be a physical capability if we had them).
	   In CHERI-MIPS this is a physical address.
bus_offset_t: offset in the bus space. This is a non-dereferenceable address in the bus space, it is intended
	     to serve the same purpose of vm_offset_t in the bus-specific code (e.g. to represent limits of a region
	     of bus space). In CHERI-MIPS this is a physical address.
bus_space_handle_t: opaque handle representing a mapped bus region. In CHERI-MIPS this is an uintptr_t,
		    the handle is the virtual address-space capability to the region of memory that maps
		    the physical bus space region.