Use the same implementation for various bcopy/memcpy variants in the purecap kernel.
Capability preserving copies:
bcopy and memcpy are now the same as bcopy_c and memcpy_c.
cheri_bcopy and cheri_memcpy are now the same as bcopy_c and memcpy_c.

Capability non-preserving copies:
bcopynocap_c and memcpynocap_c are unchanged.

Make the default kernel bcopy and memcpy the same as memcpy_c and bcopy_c.