Fix alignment check in cheri_memcpy_c variants.
When we decide whether to perform an aligned copy, we only check whether the unaligned copy size matches mod dword or capability_size.
We do not check whether there are enough bytes to copy in an aligned manner: if we are copying 8 bytes total that cross dword boundary,
we first copy N bytes up to the first dword aligned boundary, then enter the dword aligned copy loop to copy the rest.
If the remaining (8 - N in the example above) bytes overflow the length of the copy, we should not copy the full dword.
This manifests itself as a length violation when invariants are enabled.

This fix checks that we have enough bytes to perform at least one iteration of the aligned copy loops.