Revert "Unborrow the thread not just on syscall entry, but on any other trap,"
For this to work, we would need to handle a situation when we enter
the kernel on a non-syscall trap (eg on an interrupt), then swap the
userspace contexts (the trapframes), and then return as a syscall.
Or the other way around.  I don't see a way to do this in a non-ugly
and maintainable manner.

This reverts commit 0bf9f8141f1e1ddde87ead9af471edf2425072b2.