Fix semop.
FreeBSD has a weird split between the semid used internally and the
semid exposed to userspace.  The latter is the former with a 16-bit
sequence number in the to bits.  Supposidly this is to prevent reuse
races.  When converting sys_semop() to kern_semop() I failed to handle
the fact that the user semid is used to access the sequence number after
then point where I stripped it off.  This resulted in all operations
returning EINVAL.