Fix badinstr value sometimes being wrong in log_trap_frame()
If we get a TLB miss, etc before the print line we will read the instruction
that caused the TLB miss instead of the actual trap instruction. We now
store badinstr and badinstr_p in struct trapframe so it will be correct
unless we get a trap during the first few instructions of trap().