Add a variant of cheri_trapframe_to_cheriframe() which strips tags.
Use this variant in fill_capregs() to ensure we don't leak capabilities
from the target process into the debugger process.

The other uses of cheri_trapframe_to_cheriframe() are used within the
context of a single process (signal delivery, etc.) so I've left them
as preserving tags so that signal can modify the saved registers.