Rework handling of arbitrary PC values for PT_CONTINUE with CheriABI.
Allow arbitrary PC values in the CheriABI syscall wrapper, but instead
change ptrace_set_pc() to fail if a PC value is used for a CheriABI
process as the target.

This permits either CheriABI or non-CheriABI debuggers to change the PC
of non-CheriABI processes and doesn't permit the PC of CheriABI processes
to be changed.  Add a comment that we could perhaps rework this in the
future to permit PC values that would be valid in the current PCC.  This
would ensure that hybrid processes with a constrained PCC would also
constrain the range of PC values that could be set via PT_CONTINUE.