Check if there is enough data before loading a dword
In the double-word copy section of bcopy_c we need to check if there
is a double-words worth of data. If not use the copy head code to
perform the final copy rather than loading before the start of the
capability.

This was found, and fixes the cheri libc/db/db_test:bsize_ffactor
test.