Update LG_QUANTUM values for mips64 and cheri128.
LG_QUANTUM is supposed to represent the base alignment for a given
platform.  mips64 uses 16-byte stack-alignment by default in LLVM, and
also in FreeBSD (though the final fix for this isn't yet merged to
CheriBSD), so it should be using 4 instead of 3.  Similarly, cheri128
uses 16-byte alignment rather than 32.

This change changes mips64 from 8 to 16, and cheri 128 from 32 to 16.