Avoid implicit CToPtr instructions in the kernel
Implicit CToPtr is really confusing because it returns 0 if the value
is not within the bounds of $ddc. I think in all of these cases we
actually want to be storing the vaddr since it would be nice to no
longer include all of userspace in the kernel $ddc.

I found these cases by adding a new warning to clang that triggers
for (most) CToPtr uses.