Use a loop over the two register files to populate syscall arguments.
The PAD_L and PAD_R macros for CheriABI are already defined to layout the
UAP structures for syscall arguments such that the structures are effectively
an array of capabilities with scalar values stored in the offset field of
the capability.  Make use of this to avoid the need for the fill_uap helper
functions and dispatch table.  Instead, iterate over the syscall arguments
using the mask to choose which register bank to pull the next argument from
(and increment that register bank's offset).