Really fix SIG_IGN, etc. definitions
The previous attempt (fe5ec980c703ef9e14222f766ea5f20d2dbab507) was
wrong because it would cause a CFromPtr to be used in the kernel.
In the kernel we need to cast to __intcap_t since we are assigning to a
capability. If we did this in hybrid userspace we would generate a
CToPtr which causes all of the SIG_* constants to evaluate to 0.