Minor cleanup of conversion from user sa_handler to kernel sa_handler
The kernel sa_handler is a capability so when we have a MIPS user pointer
we have to either derive it from the user $pcc or store it as an untagged
capability (e.g. for SIG_IGN). The previous check tested the values of
SIG_IGN and SIG_DFL. Replace the if statements with a single
is_magic_sighander_constant() that just checks if it is a small integer
to also handle SIG_HOLD (and potential future additions.