Add a I_REALLY_MEAN_NO_CLEAN option that skips the delete-old step
Since this forks a shell that spawns a few thousand processes (sequentially!)
this step takes a few seconds and does nothing useful in 99.99% of cases.
In the remaining 0.01% of cases I can just delete the whole objtree which
is what cheribuild.py --clean does.

This saves 8 seconds for every build.