Don't ignore fatal trap signals in init if they are SIG_DFL
This ensure that we don't end up in the annoying inifinite loop when init
is miscompiled or we break csu bits.
I did this by testing for SIGPROP_SBUNWIND since this handles all the
signals that I think should terminate init during boot. Now we get a
"panic: Going nowhere without my init!" instead of an infinite loop of
CHERI traps when init is broken.