Fix purecap kernel busdma function that flushes the cache lines that contain data involved in a DMA operation.
When the buffer is not cache line size aligned, we can not access the rest of the cache line because we do not
have a capability for it. This is a poor condition for the purecap kernel, ideally all buffers involved in
DMA transfers should be cache line size aligned.
For now we access the rest of the cache lines using the global kernel root capability.