Rework the handling of the addr argument.
Allow existing mapping to be replaced using the MAP_FIXED flag IFF the
addr argument is a valid capability with the VMMAP user permission. In
this case, the new capability is derived from the passed capability. In
all other cases, the new capability is derived from the per-thread mmap
capability.

If MAP_FIXED specified and addr does not meet the above requirements,
then MAP_EXCL is implied to prevent changing page contents without
permission.

This allows MAP_FIXED in many more circumstances.  I belive MAP_FIXED
except within an existing reservation is a bad idea, but it's mostly
harmless with the conventional process model.