Add the ability to append suffixes to types based on annotations.
When the argument points to a type which is changed in this ABI (as
indicated by abi_flags), append abi_type_suffix to the struct or union
name.

Newly supported abi_flags values are: long_size, pointer_size, and
time_t_size which correspond to the _Contains annotations described
below.

The _Contains family of SAL-like annotations indicates if a pointer
points to an object containing types that require translation.  Three
types are supported:
 - _long_ is a catch-all for long, size_t etc
 - _ptr_ is for pointers
 - _timet_ is for time_t