The kernel interface to fcntl(2) takes an intptr_t.
fcntl() is declared as:

int fcntl(int fd, int cmd, ...);

with the variadic argument being an optional third argument.
Historically, this argument was a long, but with the F_GETLK, F_SETLK,
and F_SETLKW commands it is a pointer so intptr_t is the most
appropriate type. This has no effect on current architectures, but on
architectures where long and intptr_t are not synonymous (e.g. CHERI),
this change is required.