Fix kernel stack unwinding end-of-function false-positive
The kernel stack unwinder assumes that any jr $ra indicates the end of the
current function. However, clang will generate code that contains jr ra
at various places inside the function. This was causing the stack unwinder
to assume that kern_execve had a stack size of zero which is clearly wrong.
After this change it successfully detects the stack size of 960 and shows
that kern_execve was called by sys_execve.