Add a global RWX allocator to rtld (for trampolines)
Currently PLT stubs and function pointer exports all use this allocator to
get RWX memory. It does not support freeing of the memory but that is fine
since we don't really support unloading of shared libraries.
This could be improved later but for now this simple bump the pointer
allocator works just fine.