rtld: Use a vector instead of uthash for function pointer export lookups
UTHash doesn't seem to work properly in purecap mode. Furthermore, it add
huge amounts of overhead to a tiny struct. For now just iterate over a
vector of pairs to find whether a function already has an export thunk.

If this ends up being slow we can always use a hash table later.